requirements.txt
# =============================================
# 캡스톤 프로젝트 - 비지도 학습 이상 탐지 시스템
# Windows + CPU 환경용
# =============================================

# PyTorch (CPU 전용) - 아래 명령어로 별도 설치 필요:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
# (requirements.txt에는 포함하지 않음)

# 핵심 라이브러리
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
tqdm>=4.65.0

# 웹 서버
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.0

# 유틸리티
pydantic>=2.0.0
